fix(prelaunch): 3 post-bundle deploy fixes surfaced during live rollout
1. JwtTokenProvider @Autowired (api-gateway):
   The new kid-aware constructor + the legacy 3-arg test ctor are two
   constructors on a @Component bean. Spring's implicit single-ctor
   autowiring only fires with exactly one constructor — with two and
   neither annotated, it falls back to looking for a no-arg default,
   doesn't find one, and the entire ApplicationContext bails at boot
   ("No default constructor found"). Annotate the production ctor.
   Same pattern as commit 8c32368 (TestCaseFetcher / AgentClient).

2. IdempotencyKey.attempts widened int → long (execution-worker):
   CRDB's INT is BIGINT-flavoured; the V4 migration creates
   `attempts INT NOT NULL DEFAULT 0` which Hibernate ddl-auto=validate
   reads as Types.BIGINT and rejects against Java `int` (Types.INTEGER).
   Same trap as Problem.java (commit dd7eb45). Widen to long.

3. Flyway baseline-on-migrate + baseline-version=0 in compose:
   When api-gateway first connects to `onlinejudge`, the DB already
   has users / problems / test_cases from the retired init.sql path,
   but no flyway_schema_history. Default Flyway behaviour refuses to
   proceed in that state. baseline-on-migrate=true creates the history
   table; baseline-version=0 ensures V1 runs (V1 uses IF NOT EXISTS,
   so it composes safely with the partial pre-existing schema).
   These env vars are wired in control-plane-compose.yml.

All three were caught in the live rollout of commit 1b82186 and
required additional rebuild+push of the api-gateway and execution-worker
images. Now committed for durability.

+++ execution-worker/src/main/java/com/onlinejudge/worker/model/IdempotencyKey.java
@@ -26,8 +26,10 @@ public class IdempotencyKey {
     /** Roadmap §2.5 — reclaim cycles. Bumped on every stale-row reclaim;
      *  once it exceeds {@code app.idempotency.max-attempts} the next claim
      *  is treated as POISON and the consumer dead-letters the submission. */
+    // CRDB INT = BIGINT; Hibernate ddl-auto=validate rejects Java `int` (INTEGER).
+    // Widen to long to match the column's actual storage width.
     @Column(nullable = false)
-    private int attempts;
+    private long attempts;
 
     @Column(name = "created_at", nullable = false)
     private Instant createdAt = Instant.now();
